Refactor package name from person to member
