+++ src/main/java/seedu/address/model/member/exceptions/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.member.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified member.
