+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -17,8 +17,8 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
-import seedu.address.model.person.Member;
+import seedu.address.model.member.NameContainsKeywordsPredicate;
+import seedu.address.model.member.Member;
 import seedu.address.testutil.EditMemberDescriptorBuilder;
 
 /**
