+++ src/test/java/seedu/address/logic/commands/DeleteCommandTest.java
@@ -17,7 +17,7 @@ import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Member;
+import seedu.address.model.member.Member;
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for
