+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -18,8 +18,8 @@ import org.junit.jupiter.api.Test;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Member;
-import seedu.address.model.person.exceptions.DuplicateMemberException;
+import seedu.address.model.member.Member;
+import seedu.address.model.member.exceptions.DuplicateMemberException;
 import seedu.address.testutil.MemberBuilder;
 
 public class AddressBookTest {
